requirements.txt
ca-certs-locater==1.0
Django==1.11.2
django-appconf==1.0.2
django-compressor==2.1.1
django-debug-toolbar==1.8
django-honeypot==0.6.0
django-libsass==0.7
django-memoize==2.1.0
future==0.16.0
httplib2==0.9.2
libsass==0.12.3
oauth2==1.9.0.post1
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.3
urllib3==1.19
git+https://github.com/TechnicalBro/python-upwork
